Refactor `k8s` module to optimize ownership traversal logic. Allows location of topmost Pod Owner
