+++ mermin/src/k8s/resource_parser.rs
@@ -1,82 +1,10 @@
-use std::{
-    collections::{HashMap, VecDeque},
-    net::{IpAddr, Ipv4Addr, Ipv6Addr},
-};
+use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};
 
 use anyhow::Result;
-use k8s_openapi::{
-    api::{
-        apps::v1::{DaemonSet, Deployment, ReplicaSet, StatefulSet},
-        batch::v1::Job,
-    },
-    apimachinery::pkg::apis::meta::v1::OwnerReference,
-};
-use kube::{Resource, ResourceExt};
-use log::warn;
 use mermin_common::{IpAddrType, PacketMeta};
 use network_types::ip::IpProto;
 
-use crate::k8s::Attributor;
-
-/// Holds metadata for a single Kubernetes object.
-#[derive(Debug, Clone)]
-#[allow(dead_code)]
-pub struct K8sObjectMeta {
-    pub kind: String,
-    pub name: String,
-    pub uid: Option<String>,
-    pub namespace: Option<String>,
-    pub labels: Option<HashMap<String, String>>,
-    pub annotations: Option<HashMap<String, String>>,
-}
-
-/// Generic implementation to convert any Kubernetes Resource into our K8sObjectMeta.
-impl<T> From<&T> for K8sObjectMeta
-where
-    T: Resource<DynamicType = ()>,
-{
-    fn from(resource: &T) -> Self {
-        Self {
-            kind: T::kind(&()).to_string(),
-            name: resource.name_any(),
-            uid: resource.uid(),
-            namespace: resource.namespace(),
-            labels: (!resource.labels().is_empty())
-                .then(|| resource.labels().clone().into_iter().collect()),
-            annotations: (!resource.annotations().is_empty())
-                .then(|| resource.annotations().clone().into_iter().collect()),
-        }
-    }
-}
-
-/// Represents the workload controllers that own other resources, like Pods.
-#[derive(Debug, Clone)]
-pub enum WorkloadOwner {
-    ReplicaSet(K8sObjectMeta),
-    Deployment(K8sObjectMeta),
-    StatefulSet(K8sObjectMeta),
-    DaemonSet(K8sObjectMeta),
-    Job(K8sObjectMeta),
-}
-
-#[derive(Debug, Clone)]
-#[allow(dead_code)]
-pub enum EnrichedInfo {
-    Pod {
-        pod: K8sObjectMeta,
-        owners: Vec<WorkloadOwner>,
-    },
-    Node {
-        node: K8sObjectMeta,
-    },
-    Service {
-        service: K8sObjectMeta,
-        backend_ips: Vec<String>,
-    },
-    EndpointSlice {
-        slice: K8sObjectMeta,
-    },
-}
+use crate::k8s::{Attributor, EnrichedInfo, K8sObjectMeta};
 
 #[derive(Debug, Default)]
 #[allow(dead_code)]
@@ -94,106 +22,14 @@ pub struct FlowSide {
     protocol: IpProto,
 }
 
-fn get_owner_from_store(
-    attributor: &Attributor,
-    owner_ref: &OwnerReference,
-    namespace: &str,
-) -> Option<(WorkloadOwner, Option<Vec<OwnerReference>>)> {
-    let name = &owner_ref.name;
-    macro_rules! find_in_store {
-        ($store_type:ty, $variant:ident) => {
-            attributor
-                .resource_store
-                .get_by_namespace::<$store_type>(namespace)
-                .iter()
-                .find(|obj| obj.name_any() == *name)
-                .map(|obj| {
-                    let meta = K8sObjectMeta::from(obj.as_ref());
-                    let next_owners = obj.meta().owner_references.clone();
-                    (WorkloadOwner::$variant(meta), next_owners)
-                })
-        };
-    }
-
-    match owner_ref.kind.as_str() {
-        "ReplicaSet" => find_in_store!(ReplicaSet, ReplicaSet),
-        "Deployment" => find_in_store!(Deployment, Deployment),
-        "StatefulSet" => find_in_store!(StatefulSet, StatefulSet),
-        "DaemonSet" => find_in_store!(DaemonSet, DaemonSet),
-        "Job" => find_in_store!(Job, Job),
-        _ => {
-            warn!(
-                "Owner lookup for kind '{}' is not implemented.",
-                owner_ref.kind
-            );
-            None
-        }
-    }
-}
-
-/// Recursively discovers the entire ownership chain for a given object.
-pub fn find_owner_chain(
-    attributor: &Attributor,
-    initial_owners: Vec<OwnerReference>,
-    initial_namespace: &str,
-    max_depth: usize,
-) -> Vec<WorkloadOwner> {
-    let mut chain = Vec::new();
-    let mut owners_to_process: VecDeque<(OwnerReference, String, usize)> = initial_owners
-        .into_iter()
-        .map(|owner| (owner, initial_namespace.to_string(), 1))
-        .collect();
-
-    while let Some((owner_ref, namespace, depth)) = owners_to_process.pop_front() {
-        if depth > max_depth {
-            continue;
-        }
-
-        if let Some((owner_obj, next_owners_opt)) =
-            get_owner_from_store(attributor, &owner_ref, &namespace)
-        {
-            if let Some(next_owners) = next_owners_opt {
-                let owner_namespace = match &owner_obj {
-                    WorkloadOwner::ReplicaSet(m) => m.namespace.clone(),
-                    WorkloadOwner::Deployment(m) => m.namespace.clone(),
-                    WorkloadOwner::StatefulSet(m) => m.namespace.clone(),
-                    WorkloadOwner::DaemonSet(m) => m.namespace.clone(),
-                    WorkloadOwner::Job(m) => m.namespace.clone(),
-                }
-                .unwrap_or(namespace);
-
-                for next_owner in next_owners {
-                    owners_to_process.push_back((next_owner, owner_namespace.clone(), depth + 1));
-                }
-            }
-            chain.push(owner_obj);
-        } else {
-            warn!(
-                "Failed to find owner {} ({}) in store for namespace {}",
-                owner_ref.name, owner_ref.kind, namespace
-            );
-        }
-    }
-    chain
-}
-
 /// Enriches a single side of a flow (source or destination) based on its IP address.
 async fn enrich_side(side: &FlowSide, attributor: &Attributor) -> Option<EnrichedInfo> {
     if let Some(pod) = attributor.get_pod_by_ip(side.ip).await {
         let pod_meta = K8sObjectMeta::from(pod.as_ref());
-        let owners = if let Some(owner_refs) = &pod.metadata.owner_references {
-            find_owner_chain(
-                attributor,
-                owner_refs.clone(),
-                &pod.namespace().unwrap_or_default(),
-                10, // Replace with conf value when available
-            )
-        } else {
-            Vec::new()
-        };
+        let owner = attributor.get_top_level_controller(&pod);
         return Some(EnrichedInfo::Pod {
             pod: pod_meta,
-            owners,
+            owner,
         });
     }
 
@@ -212,7 +48,7 @@ async fn enrich_side(side: &FlowSide, attributor: &Attributor) -> Option<Enriche
 
         return Some(EnrichedInfo::Service {
             service: service_meta,
-            backend_ips: backend_ips.unwrap(),
+            backend_ips: backend_ips.unwrap_or_default(),
         });
     }
